Move ValidatedInputControl into with-validation folder
