+++ packages/ui/src/form/index.ts
@@ -3,4 +3,4 @@ export * from './primitives';
 export * from './input-control';
 export * from './select-control';
 export * from './textarea-control';
-export * from './validated-input-control';
+export * from './with-validation';
