+++ packages/ui/src/form/with-validation/index.ts
@@ -1,0 +1,1 @@
+export * from './validated-input-control';
