+++ packages/ui/src/form/with-validation/validated-input-control/validated-input-control.tsx
@@ -1,7 +1,7 @@
 import { forwardRef, useRef } from '@wordpress/element';
 import { useMergeRefs } from '@wordpress/compose';
-import { ControlWithError } from '../primitives/control-with-error';
-import { InputControl } from '../input-control';
+import { ControlWithError } from '../../primitives/control-with-error';
+import { InputControl } from '../../input-control';
 import type { ValidatedInputControlProps } from './types';
 
 /**
